Add support for passive buzzer
Passive buzzer support - for both passive and active buzzer I highly recommend using a transistor or FET driver (as shown in the updated wiring diagrams) This means an extra wire is necessary from the 5v pad.
For an N channel Mosfet. Source is Ground, Drain is the negative terminal of the buzzer and Gate is the buzzer pin D12. For an NPN BJT Transistor, Emitter is ground, Collector is the negative terminal of the buzzer and Base is the buzzer pin D12.